Data and Plotting Update
Fixed pandas dataframe duplication by adding the inplace=True to the dataframe definition.  Also got Plotly working and setup with the same scale as the original paper.  Still cannot get the plot to look like the plot in the paper.  Have tried changing the the initial and final states and updated the dataframe to match the paper.